Ajout d'une fonction test advance() pour QTrees. Test avec des arbres positionnés aléatoirement.

+++ GPA789-Projet/QTrees.h
@@ -22,9 +22,9 @@ public:
 
 
 private:
-	int mHeight;
-	int mTrunkRadius;
-	int mLeafRadius;
+	double mHeight;
+	double mTrunkRadius;
+	double mLeafRadius;
 	int mDryness;
 	bool mGotHit;
